requirements.txt
pytest
python-dotenv
openai